FEAT(Core): Example with two plugins
Testing to see how abi_stable::InterfaceType handles an instance created in one dynamic library and downcast in another - it works!

On another note, while crate-type="cdylib" works with abi_stable, it does not allow pulling symbols from the first plugin to the second, dependent plugin. This is very important for our use-case, because most plugins _will_ depend on others. Fortunately "dylib" seems to handle this fine for now