FEAT(Core): PoC data flow between dependent plugins
This commit had to rearrange the driver for testing, turns out load_from_file caches the loaded library, keyed by interface but not by filename, which means previously the same library was called twice.

A patch to the abi_stable crate was necessary to allow downcasting of data that passed the FFI boundary. I still do not know if the cast is safe, but the code works even with -Z randomize-layout.

Theoretically the StringBuilder type from plugin1 built as a cdylib and the StringBuilder from plugin2 with a plugin1 dependency could be laid out differently in memory, because they're compiled in different invocations of `rustc`, which is why we could add a bound to the downcast function to also be StableAbi (and repr(C)). I do not know if pinning layout this way is enough.