feature/abi_stable_poc
* FEAT(Core): Example with two plugins

Testing to see how abi_stable::InterfaceType handles an instance created in one dynamic library and downcast in another - it works!

On another note, while crate-type="cdylib" works with abi_stable, it does not allow pulling symbols from the first plugin to the second, dependent plugin. This is very important for our use-case, because most plugins _will_ depend on others. Fortunately "dylib" seems to handle this fine for now

* FEAT(Core): Rollback to cdylib

Turns out abi_stable only supports crate-type="cdylib". Will document this, the solutions and drawbacks in the PR and a following ADR

* FEAT(Core): PoC data flow between dependent plugins

This commit had to rearrange the driver for testing, turns out load_from_file caches the loaded library, keyed by interface but not by filename, which means previously the same library was called twice.

A patch to the abi_stable crate was necessary to allow downcasting of data that passed the FFI boundary. I still do not know if the cast is safe, but the code works even with -Z randomize-layout.

Theoretically the StringBuilder type from plugin1 built as a cdylib and the StringBuilder from plugin2 with a plugin1 dependency could be laid out differently in memory, because they're compiled in different invocations of `rustc`, which is why we could add a bound to the downcast function to also be StableAbi (and repr(C)). I do not know if pinning layout this way is enough.

* FEAT(Core): Implement main calling loop

Very rough kernel code is ready and works, but there is still work to be done, e.g.:
- Extract the toposort to a separate file and write unit tests for it
- Implement generic user parameter parsing and --help that extracts help messages from plugins
- Better error handling
- Plugin loading should ignore non-library files
- Plugin loading should be OS-agnostic or at least work for Linux (tested on MacOS)